+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/config/cache/CoreCacheConfiguration.java
@@ -1,5 +1,9 @@
 package eu.bcvsolutions.idm.core.config.cache;
 
+import java.time.Duration;
+import java.util.HashMap;
+import java.util.UUID;
+
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
@@ -8,12 +12,14 @@ import eu.bcvsolutions.idm.core.api.config.cache.IdMCacheConfiguration;
 import eu.bcvsolutions.idm.core.api.config.cache.LocalIdMCacheConfiguration;
 import eu.bcvsolutions.idm.core.model.service.impl.DefaultConfigurationService;
 import eu.bcvsolutions.idm.core.model.service.impl.DefaultGroovyScriptService;
+import eu.bcvsolutions.idm.core.security.api.service.AuthorizationManager;
 import groovy.lang.Script;
 
 /**
  * Configuration of caches used by core module
  *
  * @author Peter Štrunc <peter.strunc@bcvsolutions.eu>
+ * @author Radek Tomiška
  */
 @Configuration
 public class CoreCacheConfiguration {
@@ -47,5 +53,38 @@ public class CoreCacheConfiguration {
 				.withValueType(Script.class)
 				.build();
 	}
+	
+	/**
+	 * Define distributed cache for {@link AuthorizationManager} - logged identity permissions for entities.
+	 *
+	 * @return permission cache
+	 * @since 10.4.1
+	 */
+	@Bean
+	@SuppressWarnings("rawtypes")
+	public IdMCacheConfiguration permissionCacheConfiguration() {
+		return DistributedIdMCacheConfiguration.<UUID, HashMap> builder()
+			.withName(AuthorizationManager.PERMISSION_CACHE_NAME)
+				.withKeyType(UUID.class)
+				.withValueType(HashMap.class)
+				.withTtl(Duration.ofMinutes(1)) // permissions are based on data structure => cache should be effective short time (one request)
+				.build();
+	}
 
+	/**
+	 * Define distributed cache for {@link AuthorizationManager} - logged identity authorization policies.
+	 *
+	 * @return autorization policy cache
+	 * @since 10.4.1
+	 */
+	@Bean
+	@SuppressWarnings("rawtypes")
+	public IdMCacheConfiguration autorizationPolicyCacheConfiguration() {
+		return DistributedIdMCacheConfiguration.<UUID, HashMap> builder()
+			.withName(AuthorizationManager.AUTHORIZATION_POLICY_CACHE_NAME)
+				.withKeyType(UUID.class)
+				.withValueType(HashMap.class)
+				.withTtl(Duration.ofHours(2)) // Depends on identity is logged out. TODO: clear cache function (keys for logged identity only).
+				.build();
+	}
 }