+++ Realization/backend/core/core-impl/src/main/java/eu/bcvsolutions/idm/core/config/cache/NullCacheManager.java
@@ -17,6 +17,7 @@ import javax.cache.spi.CachingProvider;
  * @author Peter Štrunc
  */
 public class NullCacheManager implements CacheManager {
+	
 	@Override
 	public CachingProvider getCachingProvider() {
 		return Caching.getCachingProvider();