+++ Realization/backend/core/core-impl/src/test/java/eu/bcvsolutions/idm/core/security/DefaultAuthorizationManagerUnitTest.java
@@ -3,24 +3,26 @@ package eu.bcvsolutions.idm.core.security;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
-import static org.mockito.Matchers.any;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
+import java.util.UUID;
 
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.springframework.context.ApplicationContext;
 
 import eu.bcvsolutions.idm.core.api.dto.IdmAuthorizationPolicyDto;
 import eu.bcvsolutions.idm.core.api.service.IdmAuthorizationPolicyService;
+import eu.bcvsolutions.idm.core.api.service.IdmCacheManager;
 import eu.bcvsolutions.idm.core.api.service.ModuleService;
 import eu.bcvsolutions.idm.core.model.entity.IdmRole;
 import eu.bcvsolutions.idm.core.security.api.domain.IdmBasePermission;
-import eu.bcvsolutions.idm.core.security.api.service.AuthorizationManager;
 import eu.bcvsolutions.idm.core.security.api.service.SecurityService;
 import eu.bcvsolutions.idm.core.security.evaluator.BasePermissionEvaluator;
 import eu.bcvsolutions.idm.core.security.service.impl.DefaultAuthorizationManager;
@@ -38,15 +40,16 @@ public class DefaultAuthorizationManagerUnitTest extends AbstractUnitTest {
 	@Mock private IdmAuthorizationPolicyService service;
 	@Mock private SecurityService securityService;
 	@Mock private ModuleService moduleService;
+	@Mock private IdmCacheManager idmCacheManager;
 	//
-	private AuthorizationManager manager;
 	private List<IdmAuthorizationPolicyDto> enabledPolicies;
 	private final BasePermissionEvaluator evaluator = new BasePermissionEvaluator();
+	//
+	@InjectMocks
+	private DefaultAuthorizationManager manager;
 	
 	@Before
 	public void init() {		
-		manager = new DefaultAuthorizationManager(context, service, securityService, moduleService);
-		//
 		enabledPolicies = new ArrayList<>();
 		//
 		IdmAuthorizationPolicyDto policyOne = new  IdmAuthorizationPolicyDto();
@@ -64,6 +67,7 @@ public class DefaultAuthorizationManagerUnitTest extends AbstractUnitTest {
 	public void testGetPermissions() {
 		when(service.getEnabledPolicies(any(), any())).thenReturn(enabledPolicies);
 		when(context.getBean(BasePermissionEvaluator.class)).thenReturn(evaluator);
+		when(securityService.getCurrentId()).thenReturn(UUID.randomUUID());
 		//
 		Set<String> basePermissions = manager.getPermissions(new IdmRole());
 		assertEquals(2, basePermissions.size());