+++ Realization/backend/core/core-impl/src/test/java/eu/bcvsolutions/idm/core/security/evaluator/BasePermissionEvaluatorUnitTest.java
@@ -20,6 +20,7 @@ import eu.bcvsolutions.idm.test.api.AbstractUnitTest;
  * @author Radek Tomiška
  *
  */
+@SuppressWarnings("deprecation")
 public class BasePermissionEvaluatorUnitTest extends AbstractUnitTest {
 	
 	private BasePermissionEvaluator evaluator = new BasePermissionEvaluator();