+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 
 
+## [10.4.1]
+
+### Administrator
+
+- [#2366](https://redmine.czechidm.com/issues/2366) - New caches were registered:
+  - ``core:authorization-policy-cache`` - Cache stores active authorization policies of currently logged user. Cache is evicted after user log out. When authorization policies configuration is changed, then user is logged out and cache is evicted (after permissions removal only). Cache expiration is 2 hour, e.g. if user forgot to log out. **Memory usage per logged user is ~500KB, when 100 policies are used (~configured) for logged user**.
+  - ``core:permission-cache`` - Cache stores permissions (for data) of currently logged user. Cache is evicted after user log out. Cache expiration is 1 minute - if data structure is changed, then permissions are actualized after this duration. When authorization policies configuration is changed, cache is evicted (completely). **Memory usage per logged user is ~4KB for every shown record** (cache expiration is 1 minute => when 1000 record details are shown in the same time, then 4MB is consumed by this cache).
+
 ## [10.4.0]
 
 ### Administrator